requirements.txt
alembic==1.15.2
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
cffi==1.17.1
cfgv==3.4.0
click==8.2.0
cryptography==45.0.2
deepdiff==8.5.0
Deprecated==1.2.18
distlib==0.3.9
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
jsondiff==2.2.1
limits==5.2.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack-python==0.5.6
mypy_extensions==1.1.0
nodeenv==1.9.1
orderly-set==5.4.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic[email]
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
redis==6.1.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
